# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SEORx API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients:
    get:
      operationId: listClients
      tags: [clients]
      summary: List all clients accessible to current user
      security:
        - bearerAuth: []
      parameters:
        - name: orgId
          in: query
          schema:
            type: string
          description: Filter by organization ID
        - name: search
          in: query
          schema:
            type: string
          description: Search by name or domain
      responses:
        "200":
          description: List of clients
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Client"
        "401":
          $ref: "#/components/responses/Unauthorized"
    post:
      operationId: createClient
      tags: [clients]
      summary: Create a new client
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateClientBody"
      responses:
        "201":
          description: Created client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /clients/{id}:
    get:
      operationId: getClient
      tags: [clients]
      summary: Get client by ID
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "200":
          description: Client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      operationId: updateClient
      tags: [clients]
      summary: Update client
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateClientBody"
      responses:
        "200":
          description: Updated client
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Client"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      operationId: deleteClient
      tags: [clients]
      summary: Delete client
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      responses:
        "204":
          description: Deleted
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /issues/{id}/approve:
    put:
      operationId: approveIssue
      tags: [approvals]
      summary: Approve an issue fix recommendation
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApproveIssueBody"
      responses:
        "200":
          description: Approved issue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditIssue"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
  /issues/{id}/dismiss:
    put:
      operationId: dismissIssue
      tags: [approvals]
      summary: Dismiss an issue
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/IdParam"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DismissIssueBody"
      responses:
        "200":
          description: Dismissed issue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuditIssue"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
...
  /dashboard/stats:
    get:
      operationId: getDashboardStats
      tags: [dashboard]
      summary: Get dashboard summary statistics
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Dashboard statistics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DashboardStats"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /dashboard/recent-audits:
    get:
      operationId: getRecentAudits
      tags: [dashboard]
      summary: Get recent audit activity
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
      responses:
        "200":
          description: Recent audits
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Audit"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /dashboard/issue-breakdown:
    get:
      operationId: getIssueBreakdown
      tags: [dashboard]
      summary: Get issue severity breakdown across all audits
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Issue breakdown by severity and category
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueBreakdown"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /dashboard/score-trends:
    get:
      operationId: getScoreTrends
      tags: [dashboard]
      summary: Get SEO score trends over time
      security:
        - bearerAuth: []
      parameters:
        - name: clientId
          in: query
          schema:
            type: string
        - name: days
          in: query
          schema:
            type: integer
            default: 30
      responses:
        "200":
          description: Score trend data points
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScoreTrend"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      schema:
        type: string
  responses:
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
...
    Client:
      type: object
      properties:
        id:
          type: string
        orgId:
          type: string
        name:
          type: string
        domain:
          type: string
        industry:
          type: string
          nullable: true
        contactEmail:
          type: string
          nullable: true
        logoUrl:
          type: string
          nullable: true
        seoScore:
          type: number
          nullable: true
        lastAuditAt:
          type: string
          format: date-time
          nullable: true
        auditCount:
          type: integer
        issueCount:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - orgId
        - name
        - domain
        - auditCount
        - issueCount
        - createdAt
        - updatedAt
    CreateClientBody:
      type: object
      properties:
        orgId:
          type: string
        name:
          type: string
        domain:
          type: string
        industry:
          type: string
        contactEmail:
          type: string
          format: email
        logoUrl:
          type: string
      required:
        - orgId
        - name
        - domain
    UpdateClientBody:
      type: object
      properties:
        name:
          type: string
        domain:
          type: string
        industry:
          type: string
        contactEmail:
          type: string
        logoUrl:
          type: string
    Audit:
      type: object
      properties:
        id:
          type: string
        clientId:
          type: string
        clientName:
          type: string
        url:
          type: string
        status:
          type: string
          enum: [pending, running, completed, failed]
        seoScore:
          type: number
          nullable: true
        issueCount:
          type: integer
        criticalCount:
          type: integer
        highCount:
          type: integer
        mediumCount:
          type: integer
        lowCount:
          type: integer
        completedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - clientId
        - clientName
        - url
        - status
        - issueCount
        - criticalCount
        - highCount
        - mediumCount
        - lowCount
        - createdAt
        - updatedAt
...
    AuditIssue:
      type: object
      properties:
        id:
          type: string
        auditId:
          type: string
        url:
          type: string
        category:
          type: string
          enum: [meta, content, performance, links, structured_data, mobile, security, crawlability]
        severity:
          type: string
          enum: [critical, high, medium, low, info]
        title:
          type: string
        description:
          type: string
        recommendation:
          type: string
          nullable: true
        aiRecommendation:
          type: string
          nullable: true
        priorityScore:
          type: number
        status:
          type: string
          enum: [open, approved, dismissed, fixed]
        approvedBy:
          type: string
          nullable: true
        approvedAt:
          type: string
          format: date-time
          nullable: true
        affectedElement:
          type: string
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - auditId
        - url
        - category
        - severity
        - title
        - description
        - priorityScore
        - status
        - createdAt
    ApproveIssueBody:
      type: object
      properties:
        notes:
          type: string
    DismissIssueBody:
      type: object
      properties:
        reason:
          type: string
...
    DashboardStats:
      type: object
      properties:
        totalClients:
          type: integer
        totalAudits:
          type: integer
        totalIssues:
          type: integer
        criticalIssues:
          type: integer
        resolvedIssues:
          type: integer
        avgSeoScore:
          type: number
        auditsThisMonth:
          type: integer
        pendingApprovals:
          type: integer
      required:
        - totalClients
        - totalAudits
        - totalIssues
        - criticalIssues
        - resolvedIssues
        - avgSeoScore
        - auditsThisMonth
        - pendingApprovals
    IssueBreakdown:
      type: object
      properties:
        bySeverity:
          type: array
          items:
            type: object
            properties:
              severity:
                type: string
              count:
                type: integer
            required:
              - severity
              - count
        byCategory:
          type: array
          items:
            type: object
            properties:
              category:
                type: string
              count:
                type: integer
            required:
              - category
              - count
      required:
        - bySeverity
        - byCategory
    ScoreTrend:
      type: object
      properties:
        date:
          type: string
          format: date
        avgScore:
          type: number
        auditCount:
          type: integer
      required:
        - date
        - avgScore
        - auditCount
...